requirements.txt
Flask
Flask-SocketIO
eventlet
paho-mqtt
flask_sqlalchemy